Second lint pass.
