+++ packages/react-native/src/index.ts
@@ -3,7 +3,7 @@
  * @license     https://opensource.org/licenses/MIT
  */
 
-import aesthetic from './aesthetic';
+import { aesthetic } from './aesthetic';
 import { createStyled } from './styles';
 
 export * from './direction';
@@ -14,20 +14,20 @@ export * from '@aesthetic/core';
 export * from '@aesthetic/core-react';
 
 export const {
-  changeDirection,
-  changeTheme,
-  configure,
-  createComponentStyles,
-  createElementStyles,
-  getActiveDirection,
-  getActiveTheme,
-  getEngine,
-  getTheme,
-  registerDefaultTheme,
-  registerTheme,
-  renderComponentStyles,
-  subscribe,
-  unsubscribe,
+	changeDirection,
+	changeTheme,
+	configure,
+	createComponentStyles,
+	createElementStyles,
+	getActiveDirection,
+	getActiveTheme,
+	getEngine,
+	getTheme,
+	registerDefaultTheme,
+	registerTheme,
+	renderComponentStyles,
+	subscribe,
+	unsubscribe,
 } = aesthetic;
 
 // Export aliases
