+++ packages/react-native/src/theme.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { View } from 'react-native';
 import { ContextualThemeProviderProps, createThemeHelpers } from '@aesthetic/core-react';
-import aesthetic from './aesthetic';
+import { aesthetic } from './aesthetic';
 
 export const { ThemeContext, ThemeProvider, useTheme, withTheme } = createThemeHelpers(aesthetic);
 
@@ -9,9 +9,9 @@ export const { ThemeContext, ThemeProvider, useTheme, withTheme } = createThemeH
  * Provides a theme for a distinct portion of the React tree.
  */
 export function ContextualThemeProvider({ children, name }: ContextualThemeProviderProps) {
-  return (
-    <ThemeContext.Provider value={aesthetic.getTheme(name)}>
-      <View>{children}</View>
-    </ThemeContext.Provider>
-  );
+	return (
+		<ThemeContext.Provider value={aesthetic.getTheme(name)}>
+			<View>{children}</View>
+		</ThemeContext.Provider>
+	);
 }
