--- packages/react-native/tests/.eslintrc
@@ -1,5 +1,0 @@
-{
-  "rules": {
-    "babel/new-cap": "off"
-  }
-}
