+++ packages/react-native/tests/helpers.ts
@@ -1,31 +1,31 @@
 import {
-  darkTheme,
-  design,
-  lightTheme,
-  setupAesthetic,
-  teardownAesthetic,
+	darkTheme,
+	design,
+	lightTheme,
+	setupAesthetic,
+	teardownAesthetic,
 } from '@aesthetic/core/test';
-import aesthetic from '../src/aesthetic';
+import { aesthetic } from '../src/aesthetic';
 
 export const dawnTheme = design.createTheme(
-  { contrast: 'normal', scheme: 'light' },
-  lightTheme.tokens,
+	{ contrast: 'normal', scheme: 'light' },
+	lightTheme.tokens,
 );
 
 export const twilightTheme = design.createTheme(
-  { contrast: 'normal', scheme: 'dark' },
-  darkTheme.tokens,
+	{ contrast: 'normal', scheme: 'dark' },
+	darkTheme.tokens,
 );
 
 export function setupAestheticReact() {
-  // Order is important here!
-  aesthetic.registerTheme('twilight', twilightTheme);
-  aesthetic.registerTheme('dawn', dawnTheme);
-  setupAesthetic(aesthetic);
+	// Order is important here!
+	aesthetic.registerTheme('twilight', twilightTheme);
+	aesthetic.registerTheme('dawn', dawnTheme);
+	setupAesthetic(aesthetic);
 }
 
 export function teardownAestheticReact() {
-  teardownAesthetic(aesthetic);
-  dawnTheme.name = '';
-  twilightTheme.name = '';
+	teardownAesthetic(aesthetic);
+	dawnTheme.name = '';
+	twilightTheme.name = '';
 }
