+++ packages/react-native/tests/withTheme.test.tsx
@@ -18,7 +18,7 @@ describe('withTheme()', () => {
   }
 
   function WrappingComponent({ children }: { children?: React.ReactNode }) {
-    return <ThemeProvider name="dawn">{children || <View />}</ThemeProvider>;
+    return <ThemeProvider name="dawn">{children ?? <View />}</ThemeProvider>;
   }
 
   it('inherits name from component `constructor.name`', () => {
