+++ packages/react/src/styles.ts
@@ -26,6 +26,7 @@ export function useCss(
 }
 
 export function createStyled<
+	// eslint-disable-next-line @typescript-eslint/no-explicit-any
 	T extends ElementType | React.ComponentType<any>,
 	V extends object = {}
 >(
