+++ packages/react/tests/withStyles.test.tsx
@@ -3,7 +3,7 @@
 import React from 'react';
 import { render } from 'rut-dom';
 import { ThemeProvider, withStyles, WithStylesWrappedProps } from '../src';
-import { ButtonProps, createStyleSheet, Wrapper } from './__mocks__/Button';
+import { ButtonProps, createStyleSheet, Wrapper } from './__fixtures__/Button';
 import { dawnTheme, setupAestheticReact, teardownAestheticReact, twilightTheme } from './helpers';
 
 describe('withStyles()', () => {
@@ -54,7 +54,7 @@ describe('withStyles()', () => {
   }
 
   function WrappingComponent({ children }: { children?: React.ReactNode }) {
-    return <ThemeProvider name="dawn">{children || <div />}</ThemeProvider>;
+    return <ThemeProvider name="dawn">{children ?? <div />}</ThemeProvider>;
   }
 
   it('inherits name from component `constructor.name`', () => {
